requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-access-policy==1.5.0
injector==0.22.0
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
